Move implementations of Allocation methods to their own file
