+++ include/caffeine/Memory/Allocation.h
@@ -131,4 +131,14 @@ public:
   void DebugPrint() const;
 };
 
+inline bool operator!(AllocationPermissions perm) {
+  return perm == AllocationPermissions::None;
+}
+
+inline AllocationPermissions operator&(AllocationPermissions lhs,
+                                       AllocationPermissions rhs) {
+  return static_cast<AllocationPermissions>(static_cast<unsigned>(lhs) &
+                                            static_cast<unsigned>(rhs));
+}
+
 } // namespace caffeine
