+++ include/caffeine/Memory/MemHeap.inl
@@ -8,43 +8,6 @@
 namespace caffeine {
 
 /***************************************************
- * Allocation                                      *
- ***************************************************/
-
-inline const OpRef& Allocation::size() const {
-  return size_;
-}
-inline OpRef& Allocation::size() {
-  return size_;
-}
-
-inline const OpRef& Allocation::data() const {
-  return data_;
-}
-inline OpRef& Allocation::data() {
-  return data_;
-}
-
-inline const OpRef& Allocation::address() const {
-  return address_;
-}
-inline OpRef& Allocation::address() {
-  return address_;
-}
-
-inline AllocationKind Allocation::kind() const {
-  return kind_;
-}
-
-inline AllocationPermissions Allocation::permissions() const {
-  return perms_;
-}
-
-inline void Allocation::permissions(AllocationPermissions p) {
-  perms_ = p;
-}
-
-/***************************************************
  * Pointer                                         *
  ***************************************************/
 
@@ -61,20 +24,6 @@ inline bool Pointer::is_resolved() const {
   //       meant to be exposed. It should be fine but if slotmap ever starts
   //       using a different key type then it'll be necessary to rework this.
   return alloc_.second != SIZE_MAX;
-}
-
-/***************************************************
- * AllocationPermissions                           *
- ***************************************************/
-
-inline bool operator!(AllocationPermissions perm) {
-  return perm == AllocationPermissions::None;
-}
-
-inline AllocationPermissions operator&(AllocationPermissions lhs,
-                                       AllocationPermissions rhs) {
-  return static_cast<AllocationPermissions>(static_cast<uint>(lhs) &
-                                            static_cast<uint>(rhs));
 }
 
 } // namespace caffeine
